mizerAgents: AI Agent Support for Mizer

mizerAgents makes it easy to set up AI coding agents to work with the mizer package for dynamic multi-species size-spectrum modelling. It bundles a mizer reference card and API documentation optimised for large language models, and deploys them into any mizer project with a single call to setup_mizer_agent().

See also

Author

Maintainer: Gustav Delius gustav.delius@york.ac.uk (ORCID) [copyright holder]

Authors:

Gustav Delius gustav.delius@york.ac.uk (ORCID) [copyright holder]